fix file_size type mismatch: Long -> Integer to match int4 DB column
V1 schema defined file_size as INTEGER (int4), not BIGINT.
Hibernate schema validation was failing on startup.
Cast file.getSize() to int in ImageService; 5MB max fits safely.

+++ backend/src/main/java/com/keepernotes/service/ImageService.java
@@ -54,7 +54,7 @@ public class ImageService {
                 .note(note)
                 .url(url)
                 .r2Key(key)
-                .fileSize(file.getSize())
+                .fileSize((int) file.getSize())
                 .build();
 
         return NoteImageResponse.from(imageRepository.save(image));
